requirements.txt
finance
requests
pandas